Add nondeterminism.
Add the ability to specify non-deterministic TMs. Listing multiple
outcomes for a particular state-input combo will create a
non-deterministic split instead of using only the latest one in the list.

Deterministic TMs should use a `.dtm` suffix, and non-deterministic TMs
should use `.ntm`. TMs with the extension `.tm` will be assumed to be
non-deterministic for compatability.